+++ packages/iov-bcp-types/src/connection.ts
@@ -69,6 +69,17 @@ export interface BcpTicker {
    * names and "Holo" and "Hydro Protocol".
    */
   readonly tokenName: string;
+
+  /**
+   * The number of fractional digits the token supports.
+   *
+   * A quantity is expressed as atomic units. 10^fractionalDigits of those
+   * atomic units make up 1 token.
+   *
+   * E.g. in Ethereum 10^18 wei are 1 ETH and from the quantity 123000000000000000000
+   * the last 18 digits are the fractional part and the rest the wole part.
+   */
+  readonly fractionalDigits: number;
 }
 
 export enum BcpTransactionState {
@@ -184,8 +195,8 @@ export interface BcpConnection {
   readonly disconnect: () => void;
   readonly chainId: () => ChainId;
   readonly height: () => Promise<number>;
-  readonly getTicker: (ticker: TokenTicker) => Promise<BcpQueryEnvelope<BcpTicker>>;
-  readonly getAllTickers: () => Promise<BcpQueryEnvelope<BcpTicker>>;
+  readonly getTicker: (ticker: TokenTicker) => Promise<BcpTicker | undefined>;
+  readonly getAllTickers: () => Promise<ReadonlyArray<BcpTicker>>;
 
   // accounts
   /**